% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{An ANOVA table function}
\usage{
anova_table(..., n_digits = 3)
}
\arguments{
\item{...}{The models to use in the ANOVA table comparison

Their values may be:
\itemize{
\item \code{asreml}
\item \code{glmmTMB}
}}

\item{n_digits}{\code{numeric}
The number of digits to round results to.}
}
\value{
\code{data.frame}
A dataframe containing all ANOVA tables. Can be used with xtable to
produce a LaTeX table.
}
\description{
This function allows you to observe the ANOVA table for multiple \pkg{ASReml}
or \CRANpkg{glmmTMB} models.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)
library(asreml)
test_data <- oats
test_data["yield2"] <- oats["yield"] * runif(nrow(oats["yield"]))
mod1 <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~ idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~ idv(units),
    data = test_data
)
mod2 <- asreml(
    fixed = yield2 ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~ idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~ idv(units),
    data = test_data
)
anova_table(mod1, mod2)

## With glmmTMB models
library(glmmTMB)
Salamanders$count2 <- runif(nrow(Salamanders), 0, 10)
mod3 <- glmmTMB( # Zero inflated model
    count ~ spp * mined + (1 | site),
    zi = ~ spp * mined,
    data = Salamanders,
    family = nbinom2
)
mod4 <- glmmTMB( # Hurdle model
    count2 ~ spp * mined + (1 | site),
    zi = ~ spp * mined,
    data = Salamanders,
    family = truncated_nbinom2
)
anova_table(mod3, mod4)
\dontshow{\}) # examplesIf}
}
