% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_variograms.R
\name{directional_variograms}
\alias{directional_variograms}
\title{A spatial variogram function for Single Site Data}
\usage{
directional_variograms(model)
}
\arguments{
\item{model}{\code{asreml}, an ASReml model with some residual structure.}
}
\value{
A \code{gstat::variogram} object.
}
\description{
This function allows you to observe directional variograms in the 0 (vertical) and 90 (horizontal) directions for gridded small-plot trial data.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ weed,
    random = ~idv(Variety),
    residual = ~ar1v(Row):id(Column),
    data = wheat
)
mod <- directional_variograms(model)
plot(mod, multipanel = FALSE)

\dontrun{
# You can also plot using ggplot if you wish
ggplot(
    mod,
    aes(x = dist, y = gamma, group = dir.hor, colour = factor(dir.hor))
) +
    geom_point() +
    geom_line() +
    facet_grid(~dir.hor, scales = "free_x")
}
\dontshow{\}) # examplesIf}
}
