% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_prediction_file.R
\name{excel_prediction_file}
\alias{excel_prediction_file}
\title{A prediction function for models with more than one variable to predict on.}
\usage{
excel_prediction_file(model, classify, filename, sed = F)
}
\arguments{
\item{model}{an ASReml model. Can be a model object, or a string referencing
a model object.}

\item{classify}{A string specifying the factors in the model to predict on.
For example "Nitrogen:Variety".}

\item{filename}{A string for the filename.}

\item{sed}{logical: should sed be included in the predictions?}
}
\description{
This function outputs the predictions for each factor for a single model.
Each factor is put on a separate excel sheet.
}
\examples{
\dontrun{
excel_prediction_file(
model,
classify = "N:V",
filename = "Yield_Prediction.xlsx"
)
}
}
\keyword{latex}
\keyword{wald}
\keyword{xtable}
