% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_table.R
\name{exploratory_table}
\alias{exploratory_table}
\title{Produces a summary table for the data with Mean, Median, SD, CV and Range.}
\usage{
exploratory_table(data, response, groupby)
}
\arguments{
\item{data}{\code{data.frame}
A data set to be summarised}

\item{response}{\code{character}
String variable name in the data of the response}

\item{groupby}{\code{character} vector
Vector of strings of the names of the grouping variables in the data}
}
\value{
\code{data.frame}
Summary table
}
\description{
This function outputs the summary statistics of the given data.
}
\examples{
exploratory_table(mtcars, response = "mpg", groupby = c("cyl", "gear"))

}
