% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotborder.R
\name{ggplotborder}
\alias{ggplotborder}
\title{Produces a plot border for an experimental design.}
\usage{
ggplotborder(deslist, nameborder = NULL)
}
\arguments{
\item{deslist}{experimental design listing.}

\item{nameborder}{name of the blocking variable. For example, "Rep".}
}
\value{
a list of the raw stats tables
}
\description{
Produces a plot border for an experimental design.
}
\examples{
\dontrun{
.data <- data.frame(A = as.factor(rep(1:5, 2)), B = rnorm(10), C = as.factor(rep(1:2, 5)))
raw_stats(.data, response = "B", C, A)
}
}
\keyword{border}
\keyword{design}
