% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_graph.R
\name{lsd_graph}
\alias{lsd_graph}
\title{Create an LSD plot from an ASReml model}
\usage{
lsd_graph(model, classify, ...)
}
\arguments{
\item{model, }{an ASReml model.}

\item{classify}{A string specifying which variables to predict and calculate
LSDs from.}

\item{...}{Arguments to pass to \code{predictPlus.asreml}}
}
\value{
return a ggplot2 object.
}
\description{
Generates a table of least significant differences (LSDs) for a given model.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = oats
)
lsd_graph(model, classify = "Variety")
\dontshow{\}) # examplesIf}
}
