% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_group.R
\name{lsd_group}
\alias{lsd_group}
\title{Group treatments using results of paired t-test}
\usage{
lsd_group(treatments, means, alpha, pvalues)
}
\arguments{
\item{treatments}{character vector of the treatment names}

\item{means}{numeric vector of the treatment means/fitted values}

\item{alpha}{numeric significant difference threshold}

\item{pvalues}{symmetric numeric matrix of pvalues calculated via pairwise
t-tests}
}
\value{
data.frame of treatments and their associated group
}
\description{
Group treatments using results of paired t-test
}
\examples{
\dontshow{if ("&"(requireNamespace("asreml", quietly = TRUE), requireNamespace("asremlPlus", quietly = TRUE))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = oats
)

pred <- asremlPlus::predictPlus.asreml(
    model,
    classify = "Variety",
    wald.tab = as.data.frame(asreml::wald(model, denDF = "algebraic")$Wald)
)

prob.matrix <- ifelse(is.na(pred$p.differences), 1, pred$p.differences)
treatments <- colnames(prob.matrix)
means <- pred$predictions$predicted.value
alpha <- 0.05

lsdmeantab <-
    lsd_group(
        treatments,
        means,
        alpha,
        prob.matrix
    )
\dontshow{\}) # examplesIf}
}
