% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_table.R
\name{lsd_table}
\alias{lsd_table}
\title{Create an LSD table from an ASReml model}
\usage{
lsd_table(model, classify, alpha = 0.05, ...)
}
\arguments{
\item{model}{The model object to calculate LSDs from.

The value may be:
\itemize{
\item \code{asreml}
\item \code{glmmTMB} (not yet implemented)
}}

\item{classify}{\code{character}
A string specifying which variables to predict and calculate LSDs from.}

\item{alpha}{xxxxxxx}

\item{...}{Arguments to pass to \code{predictPlus.asreml}}
}
\value{
A \code{data.frame} with the LSD values.
}
\description{
Generates a table of least significant differences (LSDs) for a given model.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~ idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~ idv(units),
    data = oats
)
lsd_table(model, classify = "Variety")
\dontshow{\}) # examplesIf}
}
