% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFE_rotate_and_grid.R
\name{ofe_make_penvs}
\alias{ofe_make_penvs}
\title{Make pseudo-environments for gridded OFE data}
\usage{
ofe_make_penvs(gridded_ofe, npe)
}
\arguments{
\item{gridded_ofe}{\code{gridded.ofe}.
A gridded OFE object generated by \code{\link[=ofe_grid_data]{ofe_grid_data()}}.}

\item{npe}{\code{numeric}.
Number of pseudo-environments to generate along the strips.}
}
\value{
\code{gridded.ofe}.
\code{gridded_ofe} with a \code{Pe} factor added to \code{gridded_data}.
}
\description{
Adds evenly-spaced pseudo-environment labels along the grid rows.
Always assess pseudo-environments visually before use.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)
library(dplyr)
library(sf)

data <- agridat::lasrosas.corn |>
filter(year == 2001) |>
rename(
  Yield = yield,
  Treatment = nf,
  Rep = rep
 ) |>
 dplyr::select(Yield, Treatment, Rep, long, lat) |>
 st_as_sf(coords = c("long", "lat"), crs = 4326) |>
 st_transform(3395)

gridded_data <- ofe_grid_data(data, -60, 80, 18, trim_ends=FALSE)
ofe_make_penvs(gridded_data, 5)
\dontshow{\}) # examplesIf}
}
\author{
Braden Thorne, \email{braden.thorne@curtin.edu.au}
}
