% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFE_rotate_and_grid.R
\name{ofe_rotate_data}
\alias{ofe_rotate_data}
\title{Rotate a georeferenced data frame}
\usage{
ofe_rotate_data(data, angle)
}
\arguments{
\item{data}{\code{data.frame}.
The dataframe to be rotated.
This should be goereferenced per the \code{sf} pacakge.}

\item{angle}{\code{numeric}.
Clockwise angle in degrees to rotate the dataframe.}
}
\value{
\code{data.frame}.
\code{data} with geometry now defined in rotated coordinates.
Contains the same columns as the input with the following additions:
\describe{
\item{x_original}{The original x-coordinates before rotation.}
\item{y_original}{The original y-coordinates before rotation.}
}
}
\description{
Rotate a provided georeferenced dataframe by a specified angle in degrees.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)
library(dplyr)
library(sf)

data <- agridat::lasrosas.corn |>
filter(year == 2001) |>
rename(
  Yield = yield,
  Treatment = nf,
  Rep = rep
 ) |>
 dplyr::select(Yield, Treatment, Rep, long, lat) |>
 st_as_sf(coords = c("long", "lat"), crs = 4326) |>
 st_transform(3395)

ofe_rotate_data(data, -60)
\dontshow{\}) # examplesIf}
}
\author{
Braden Thorne, \email{braden.thorne@curtin.edu.au}
}
