% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detect.R
\name{outlier_detect}
\alias{outlier_detect}
\title{A Standard Residual to detect outliers using aom}
\usage{
outlier_detect(model, threshold = 3.5, printThres = 2)
}
\arguments{
\item{model, }{an asreml model object}

\item{threshold, }{threshold at which the standard residual is cut to default 3.5}

\item{printThres, }{The number of closer to threshold values to print default 2, max length of data}
}
\description{
This function outputs the any potential outliers in a asreml model
}
\examples{
\dontrun{
outlier_detect(mod1, printThres = 5)
}
}
\keyword{Residuals}
\keyword{Standard}
