% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_table.R
\name{outlier_table}
\alias{outlier_table}
\title{An outlier Function}
\usage{
outlier_table(model, cutoff = 3.5, factors = NULL, data = NULL)
}
\arguments{
\item{model}{an ASReml Model.}

\item{cutoff}{the point beyond which points are considered as Outliers}

\item{factors}{the treatment factors used in the model}

\item{data}{the data used in the Model. Can be found automatically from the model.}
}
\description{
This function produces a table of the outliers as well the other points from that same treatment configuration.
}
\examples{
\dontrun{
outlier_table(model)
}
}
\keyword{Outlier}
