% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sagi.R
\name{plot_sagi}
\alias{plot_sagi}
\title{A plotting function for Single-Site and MET Analysis}
\usage{
plot_sagi(model, stndResid = TRUE)
}
\arguments{
\item{model, }{an ASReml model.}

\item{stndResid}{Use standardised residuals?}
}
\description{
This function allows you to observe outliers more easily.
}
\examples{
\dontrun{
plot_sagi(model)
}
}
\keyword{Outlier}
