% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_se_avsed.R
\name{pred_se_avsed_lsd}
\alias{pred_se_avsed_lsd}
\title{A function for Predictions, SE, AVSED and possibly LSD}
\usage{
pred_se_avsed_lsd(model, classify, LSD = FALSE, sed = F)
}
\arguments{
\item{model, }{an ASReml model.}

\item{classify, }{A string specifying the variables to consider.}

\item{LSD}{logical. Do you include the mean LSD value?}

\item{sed}{logical. Do you include the sed matrix?}
}
\description{
This function needs the library asreml, tidyr and dplyr
}
\examples{
\dontrun{pred_se_avsed_lsd(model = mod1, classify = "Nitrogen", LSD = T)}
}
\keyword{Predictions}
