% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_2D_Vario.R
\name{spatial_2D_variogram}
\alias{spatial_2D_variogram}
\title{A spatial variogram function for Single Site Data}
\usage{
spatial_2D_variogram(model, tol = 0)
}
\arguments{
\item{model}{\code{asreml}, an ASReml model with some residual structure.}

\item{tol}{\code{numeric}, tolerance for each angle that is allowed.}
}
\value{
A \code{gstat::variogram} object.
}
\description{
This function allows you to observe outliers more easily
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- wheat1.asr <- asreml(
    fixed = yield ~ weed,
    random = ~idv(Variety),
    residual = ~ar1v(Row):id(Column),
    data = wheat
)
spatial_2D_variogram(model)
\dontshow{\}) # examplesIf}
}
