% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ar1.R
\name{spatial_ar1}
\alias{spatial_ar1}
\title{A spatial variogram function for Single Site Data}
\usage{
spatial_ar1(model)
}
\arguments{
\item{model, }{an ASReml model.}
}
\description{
This function needs the library stringr, asreml and ggplot2
}
\examples{
\dontrun{
spatial_ar1(model)
}
}
\keyword{spatial}
