% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\name{updater}
\alias{updater}
\title{An Updater Function for Asreml}
\usage{
updater(model, niter = 10)
}
\arguments{
\item{model}{an ASReml Model.}

\item{niter}{Number of iterations to run before stopping.}
}
\description{
This function allows you to quickly update an Asreml model multiple times. It will stop if it hasn't converged after niter iterations.
}
\examples{
\dontrun{
model <- updater(model)
}

}
\keyword{Outlier}
