% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{An ANOVA table Function}
\usage{
anova_table(..., n_digits = 3)
}
\arguments{
\item{...}{\code{asreml}, any number of asreml models.}

\item{n_digits}{\code{numeric}, number of digits to round results to.}
}
\value{
a dataframe containing all anova tables. Can be used with xtable to
produce a LaTeX table.
}
\description{
This function allows you to observe the anova table for multiple ASReml
or glmmTMB models.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(CBADASReml)
library(asreml)
test_data <- oats
test_data["yield2"] <- oats["yield"] * runif(nrow(oats["yield"]))
mod1 <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = test_data
)
mod2 <- asreml(
    fixed = yield2 ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = test_data
)
anova_table(mod1, mod2)
\dontshow{\}) # examplesIf}
}
