% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBADASReml-package.R
\docType{package}
\name{CBADASReml-package}
\alias{CBADASReml}
\alias{CBADASReml-package}
\title{CBADASReml: Provides helper functions for ASReml}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides helper functions for ASReml, particularly for use with small-plot trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagi-aus.github.io/CBADASReml/}
}

}
\author{
\strong{Maintainer}: Kai Bagley \email{kai.bagley@curtin.edu.au} (\href{https://orcid.org/0009-0004-6579-6959}{ORCID})

Authors:
\itemize{
  \item Jordan Brown \email{jordan.brown@curtin.edu.au} (\href{https://orcid.org/0009-0008-5014-0593}{ORCID})
  \item Matthew Nguyen \email{matthew.nguyen@curtin.edu.au} (\href{https://orcid.org/0009-0006-6773-8263}{ORCID})
  \item Adam H. Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

Other contributors:
\itemize{
  \item Curtin University (http://www.curtin.edu.au/, ROR: 02n415q13) [funder, copyright holder]
  \item Grains Research and Development Corporation (Project: GRDC Project CUR2210-005OPX (AAGI-CU), ROR: 02xwr1996) [funder, copyright holder]
}

}
\keyword{internal}
