% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBADASReml-package.R
\docType{package}
\name{CBADASReml-package}
\alias{CBADASReml}
\alias{CBADASReml-package}
\title{CBADASReml: Provides helper functions for ASReml}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides helper functions for ASReml, particularly for use with small-plot trials.
}
\author{
\strong{Maintainer}: Kai Bagley \email{kai.bagley@curtin.edu.au} (\href{https://orcid.org/0009-0004-6579-6959}{ORCID})

Authors:
\itemize{
  \item Jordan jordan.brown@curtin.edu.au Brown (\href{https://orcid.org/0009-0008-5014-0593}{ORCID})
  \item Adam H. adamhsparks@gmail.com Sparks (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

Other contributors:
\itemize{
  \item Curtin University of Technology (\href{https://ror.org/02n415q13}{ROR}) (http://www.curtin.edu.au/) [funder, copyright holder]
  \item Grains Research and Development Corporation (\href{https://ror.org/02xwr1996}{ROR}) (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
