% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBADASReml-package.R
\docType{package}
\name{CBADASReml-package}
\alias{CBADASReml}
\alias{CBADASReml-package}
\title{CBADASReml: Provides helper functions for ASReml}
\description{
This package provides helper functions for ASReml, particularly for use with small-plot trials.
}
\author{
\strong{Maintainer}: Kai Bagley \email{kai.bagley@curtin.edu.au} (\href{https://orcid.org/0009-0004-6579-6959}{ORCID})

Authors:
\itemize{
  \item Jordan jordan.brown@curtin.edu.au Brown (\href{https://orcid.org/0009-0008-5014-0593}{ORCID})
  \item Adam H. Sparks \email{adam.sparks@curtin.edu.au}
}

Other contributors:
\itemize{
  \item Curtin University [copyright holder]
  \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
