% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_stats.R
\name{exploratory_table}
\alias{exploratory_table}
\title{Produces a summary table for the data with Mean, Median, SD, CV and Range.}
\usage{
exploratory_table(data, response, decimal_places = 3, ...)
}
\arguments{
\item{data, }{a data set to be summarised}

\item{response, }{variable (without quotes) to get summary statistics of.}

\item{decimal_places, }{number of decimal places to show}

\item{...}{variables (without quotes) to group and arrange the data in.}
}
\value{
a data.frame
}
\description{
This function outputs the summary statistics of the given data
}
\examples{
\dontrun{
exploratory_table(mtcars, disp, cyl, gear)
}
}
\keyword{Raw}
\keyword{statistics}
