% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_table.R
\name{exploratory_table}
\alias{exploratory_table}
\title{Produces a summary table for the data with Mean, Median, SD, CV and Range.}
\usage{
exploratory_table(data, response, groupby)
}
\arguments{
\item{data}{a data set to be summarised}

\item{response}{string variable name in the data of the response}

\item{groupby}{vector of strings of the names of the grouping variables in
the data}
}
\value{
data.frame
}
\description{
This function outputs the summary statistics of the given data
}
\examples{
exploratory_table(mtcars, response = "mpg", groupby = c("cyl", "gear"))
}
