% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_graph.R
\name{lsd_graph}
\alias{lsd_graph}
\title{Create an LSD graph from an ASReml model}
\usage{
lsd_graph(model, classify, ...)
}
\arguments{
\item{model}{The model to generate LSDs from.

The value may be:
\itemize{
\item \code{asreml}
\item \code{glmmTMB} (not yet implemented)
}}

\item{classify}{\code{character} vector
A string specifying which variables to predict and calculate LSDs from.

For \code{asreml} models, it is passed to \code{classify} argument of
\code{predictPlus.asreml}.}

\item{...}{Arguments to pass to \code{predictPlus.asreml}}
}
\value{
\code{ggplot} object
Returns a ggplot2 plot of the LSDs.
}
\description{
Generates a table of least significant differences (LSDs) for a given model.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = oats
)
lsd_graph(model, classify = "Variety")
\dontshow{\}) # examplesIf}
}
