% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_group.R
\name{lsd_group}
\alias{lsd_group}
\title{Group treatments using results of paired t-test}
\usage{
lsd_group(treatments, means, alpha, pvalues)
}
\arguments{
\item{treatments}{\code{character} vector
Character vector of the treatment names}

\item{means}{\code{numeric}
Vector of the treatment means/fitted values}

\item{alpha}{\code{numeric}
Significant difference threshold}

\item{pvalues}{\code{matrix} of \code{numeric}
Matrix of pvalues calculated via pairwise t-tests}
}
\value{
data.frame
Dataframe of each treatment and their associated LSD group
}
\description{
Intended as a replacement of the agricolae::orderPvalue function, but with
(maybe) a better algorithm.
}
\examples{
\dontshow{if ("&"(requireNamespace("asreml", quietly = TRUE), requireNamespace("asremlPlus", quietly = TRUE))) withAutoprint(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = oats
)

pred <- asremlPlus::predictPlus.asreml(
    model,
    classify = "Variety",
    wald.tab = as.data.frame(asreml::wald(model, denDF = "algebraic")$Wald)
)

prob.matrix <- ifelse(is.na(pred$p.differences), 1, pred$p.differences)
treatments <- colnames(prob.matrix)
means <- pred$predictions$predicted.value
alpha <- 0.05

lsd_group(
    treatments,
    means,
    alpha,
    prob.matrix
)
\dontshow{\}) # examplesIf}
}
