% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFE_rotate_and_grid.R
\name{ofe_grid_data}
\alias{ofe_grid_data}
\title{Grid data ready for ASReml analysis}
\usage{
ofe_grid_data(data_in, rotation_angle, nrow, npe, ncol, trim_ends = TRUE)
}
\arguments{
\item{data_in}{\code{data.frame}.
The dataframe to be rotated.
This should be georeferenced per the \CRANpkg{sf} package.

We assume there are only three column of actual interest in the dataframe,
which are:
\describe{
\item{Yield}{The observed yield values.}
\item{Treatment}{The treatment factor to assess.}
\item{Rep}{The blocking structure to be preserved}
}}

\item{rotation_angle}{\code{numeric}.
Clockwise angle in degrees to rotate the dataframe.}

\item{nrow}{\code{numeric}.
The number of rows in the final grid.
This should be as large as possible without introducing spurious NA values.}

\item{npe}{\code{numeric}.
Number of pseudo-environments to be generated along the strips.
PEs will be evenly spaced, so may not be appropriate in all cases.
Always assess PEs visually before use.}

\item{ncol}{\code{numeric}.
The actual number of strips in the data.
This should be counted/confirmed before hand.
If there is a gap between strips, count the number of strips
in this gap when determining this number.}

\item{trim_ends}{\code{bool}.
Boolean determining whether the gridded data should be trimmed
to remove NAs occurring at the ends of columns.
Defaults to FALSE.}
}
\value{
\code{gridded.ofe}.
\code{list} containing the following items:
\describe{
\item{gridded_data \code{data.frame}}{
The georeferenced gridded data.
}
\item{data_original \code{data.frame}}{
The input data with additional columns for visualising the
rotation process.
}
}
}
\description{
Optimally rotate and grid a georeferenced dataframe
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)

data <- agridat::lasrosas.corn |>
filter(year == 2001) |>
rename(
  Yield = yield,
  Treatment = nf,
  Rep = rep
 ) |>
 dplyr::select(Yield, Treatment, Rep, long, lat) |>
 st_as_sf(coords = c("long", "lat"), crs = 4326) |>
 st_transform(3395)

ofe_grid_data(data, -60, 80, 5, 18, trim_ends=FALSE)
\dontshow{\}) # examplesIf}
}
\author{
Braden Thorne, \email{braden.thorne@curtin.edu.au}
}
