% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_summary.R
\name{outlier_summary}
\alias{outlier_summary}
\title{Provides a summary of the outliers present in the asreml model.
Gives context to the outliers by showing the responses for the same factor
combinations as the outliers}
\usage{
outlier_summary(model, cutoff = 3.5)
}
\arguments{
\item{model}{an ASReml Model.}

\item{cutoff}{the point beyond which points are considered as Outliers}
}
\value{
prints to the console the number of outliers and two tables (one with just the outliers, one giving the context relevant to same factor combinations).
}
\description{
Provides a summary of the outliers present in the asreml model.
Gives context to the outliers by showing the responses for the same factor
combinations as the outliers
}
\examples{

library(asreml)
model <- asreml(
    fixed = weight ~ littersize + Dose + Sex + Dose:Sex,
    random = ~idv(Dam),
    residual = ~units,
    data = rats
)

outlier_summary(model)

}
\keyword{Outlier}
