% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_summary.R
\name{outlier_summary}
\alias{outlier_summary}
\title{Detect outliers for small-plot trial analysis}
\usage{
outlier_summary(model, cutoff = 3.5)
}
\arguments{
\item{model}{The model object to detect outliers in.

The value may be:
\itemize{
\item \code{asreml}
\item \code{glmmTMB} (not yet implemented)
}}

\item{cutoff}{\code{numeric}
The magnitude of a point's residual to be considered an outlier.}
}
\value{
\code{NULL}
Prints:
\itemize{
\item The number of outliers.
\item A table of the outliers, if any.
\item A table of relevant context to some factor combinations, if there are
outliers.
}
}
\description{
Provides a summary of the outliers present in the \link[asreml]{asreml}
model. Gives context to the outliers by showing the responses for the same
factor combinations as the outliers
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = weight ~ littersize + Dose + Sex + Dose:Sex,
    random = ~idv(Dam),
    residual = ~units,
    data = rats
)
outlier_summary(model)
\dontshow{\}) # examplesIf}
}
