% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFE_rotate_and_grid.R
\name{plot_gridded_ofe}
\alias{plot_gridded_ofe}
\title{Plot diagnostics of gridded OFE}
\usage{
plot_gridded_ofe(gridded_ofe)
}
\arguments{
\item{gridded_ofe}{\code{gridded.ofe}.
\code{list} generated by \code{ofe_grid_data} containing the following items:
\describe{
\item{gridded_data \code{data.frame}}{
The georeferenced gridded data.
}
\item{data_original \code{data.frame}}{
The input data with additional columns for visualising the
rotation process.
}
}}
}
\description{
Returns a plot with the following attributes:
\describe{
\item{Original Data}{
TOP LEFT.
The raw data in the projected space.
}
\item{Rough Rotation}{
TOP RIGHT
Rotated original data with alternative strips highlighted.
If each strip is not a single colour, the rotation is insufficient.
In practice only the first strip needs to be identified correctly,
however it is best to ensure all strips are identified.
}
\item{Gridded Data (Raw)}{
BOTTOM LEFT
The accurate rotated, gridded data.
Strips should be vertically aligned.
NAs will display in red.
The number of NAs should be relatively small, concentrated around the
edges where strips do not begin at the same point (if ends were trimmed
these should not be present) and in holes where data is missing.
Red horizontal lines indicate that too many rows have been specified.
Large NA concentrations in all four corners indicates the rotation
has likely failed.
}
\item{Gridded Data (Projected)}{
BOTTOM RIGHT
The gridded data projected back to the original data space.
This should look similar to the original data - if it does not the
process has likely failed and the other plots should be inspected to
establish why.
}
}
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)

data <- agridat::lasrosas.corn |>
filter(year == 2001) |>
rename(
  Yield = yield,
  Treatment = nf,
  Rep = rep
 ) |>
 dplyr::select(Yield, Treatment, Rep, long, lat) |>
 st_as_sf(coords = c("long", "lat"), crs = 4326) |>
 st_transform(3395)

gridded_data <- ofe_grid_data(data, -60, 80, 18, trim_ends=FALSE)
plot_gridded_ofe(gridded_data)
\dontshow{\}) # examplesIf}
}
\author{
Braden Thorne, \email{braden.thorne@curtin.edu.au}
}
