% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_table.R
\name{pred_table}
\alias{pred_table}
\title{Generate prediction table from model}
\usage{
pred_table(
  mod,
  classify,
  link_fun = "identity",
  tmb_component = "cond",
  tmb_type = "response",
  factor_combine = TRUE,
  trt_col_label = "Treatment"
)
}
\arguments{
\item{mod}{The model object to get predictions from.

The value may be:
\itemize{
\item \code{asreml}
\item \code{glmmTBM}
}}

\item{classify}{\code{character} vector.
Variable for which predictions will be made.

For \code{\link[asreml:asreml]{asreml::asreml()}} models, it is passed to \code{classify} argument of
\code{\link[asremlPlus:predictPlus.asreml]{asremlPlus::predictPlus.asreml()}}.
For \CRANpkg{glmmTMB} models it is used in the \code{specs} argument of
\code{\link[emmeans:emmeans]{emmeans::emmeans()}}}

\item{link_fun}{\code{character} vector.
Specifies the transformation function to apply over the predictions.
Only applies for \code{\link[asreml:asreml]{asreml::asreml()}} objects.

The value may be:
\itemize{
\item \code{"identity"} (default)
\item \code{"log"}
\item \code{"inverse"}
\item \code{"sqrt"}
\item \code{"logit"}
\item \code{"probit"}
\item \code{"cloglog"}
}}

\item{tmb_component}{\code{character} vector.
Specifies the component of the \CRANpkg{glmmTMB} model from which to get
predictions.

The value may be:
\itemize{
\item \code{"cond"} (default)
\item \code{"zi"}
\item \code{"cmean"}
\item \code{"response"}
}}

\item{tmb_type}{\code{character} vector.
Specifies  the prediction type for \CRANpkg{glmmTMB} models.
\itemize{
\item Only \code{"response"} is supported at the moment.
}}

\item{factor_combine}{Logical
Whether or not to combine the factors in \code{classify} into a single
column in the output table. If \code{TRUE} (default), the factor levels
are concatenated and labelled using \code{trt_col_label}.}

\item{trt_col_label}{\code{character} vector.
Specifies the label for the combined factor column when
\code{factor_combine} is \code{TRUE}. Defaults to \code{"Treatment"}}
}
\value{
\code{data.frame}.
Contains the predicted means, standard errors, and confidence intervals
for the specified variables. Includes the following columns:
\describe{
\item{Treatment}{The combined factor levels, if \code{factor_combine} is
\code{TRUE}.}
\item{Mean}{The predicted mean values.}
\item{Standard Error}{The standard errors of the predicted means.}
\item{Lower CL}{The lower confidence limits.}
\item{Upper CL}{The upper confidence limits.}
}
If \code{factor_combine} is \code{FALSE}, the factors specified in
\code{classify} remain as separate columns.
}
\description{
Generate a table of predicted values and CIs from an
\link[asreml:asreml]{asreml} or \link[glmmTMB:glmmTMB]{glmmTMB} model, for
the specified variables \code{asreml} or \code{glmmTMB} object.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(CBADASReml)
library(asreml)
library(glmmTMB)
mod1 <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~ idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~ idv(units),
    data = oats
)
# Zero inflated model
mod2 <- glmmTMB(
    count ~ spp * mined + (1|site),
    zi = ~ spp * mined,
    data = Salamanders,
    family = nbinom2
)
pred_table(mod1, classify = "Variety")
\dontshow{\}) # examplesIf}
}
\author{
Matthew Nguyen, \email{matthew.nguyen@curtin.edu.au}
}
