% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_table.R
\name{pred_table}
\alias{pred_table}
\title{Generate Prediction Table from Model}
\usage{
pred_table(
  mod,
  classify,
  link_fun = "identity",
  tmb_component = "cond",
  tmb_type = "response",
  factor_combine = TRUE,
  trt_col_label = "Treatment"
)
}
\arguments{
\item{mod}{The model object to get predictions from.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The value may be:
* \code{asreml}
* \code{glmmTBM}
}\if{html}{\out{</div>}}}

\item{classify}{\code{character} vector.
Variable for which predictions will be made.

\if{html}{\out{<div class="sourceCode">}}\preformatted{For \code{asreml} models, it is passed to \code{classify} argument of
\code{predictPlus.asreml}.
For \code{glmmTMB} models it is used in the \code{specs} argument of
\code{\link[emmeans]{emmeans}}
}\if{html}{\out{</div>}}}

\item{link_fun}{\code{character} vector.
Specifies the transformation function to apply over the predictions.
Only applies for \code{asreml} objects.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The value may be:
* \code{"identity"} (default)
* \code{"log"}
* \code{"inverse"}
* \code{"sqrt"}
* \code{"logit"}
* \code{"probit"}
* \code{"cloglog"}
}\if{html}{\out{</div>}}}

\item{tmb_component}{\code{character} vector.
Specifies the component of the \code{glmmTMB} model from which to get
predictions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The value may be:
* \code{"cond"} (default)
* \code{"zi"}
* \code{"cmean"}
* \code{"response"}
}\if{html}{\out{</div>}}}

\item{tmb_type}{\code{character} vector.
Specifies  the prediction type for \code{glmmTMB} models.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Only \code{"response"} is supported at the moment.
}\if{html}{\out{</div>}}}

\item{factor_combine}{Logical
Whether or not to combine the factors in \code{classify} into a single
column in the output table. If \code{TRUE} (default), the factor levels
are concatenated and labelled using \code{trt_col_label}.}

\item{trt_col_label}{\code{character} vector.
Specifies the label for the combined factor column when
\code{factor_combine} is \code{TRUE}. Defaults to \code{"Treatment"}}
}
\value{
\code{data.frame}.
Contains the predicted means, standard errors, and confidence intervals
for the specified variables. Includes the following columns:
\describe{
\item{Treatment}{The combined factor levels, if \code{factor_combine} is \code{TRUE}.}
\item{Mean}{The predicted mean values.}
\item{Standard Error}{The standard errors of the predicted means.}
\item{Lower CL}{The lower confidence limits.}
\item{Upper CL}{The upper confidence limits.}
}
If \code{factor_combine} is \code{FALSE}, the factors specified in
\code{classify} remain as separate columns.
}
\description{
Generater a table of predicted values and CIs from an
\code{\link[asreml]{asreml}} or \code{\link[glmmTMB]{glmmTMB}} model, for
the specified variables \code{asreml} or \code{glmmTMB} object. Created by Matthew
Nguyen.
}
