% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_table.R
\name{pred_table}
\alias{pred_table}
\title{A function that returns tablulated predictions from a model}
\usage{
pred_table(
  mod,
  classify,
  link_fun = "identity",
  tmb_component = "cond",
  tmb_type = "response",
  factor_combine = TRUE,
  trt_col_label = "Treatment"
)
}
\arguments{
\item{mod}{The model object to get predictions from.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The value may be:
* `asreml`
* `glmmTBM`
}\if{html}{\out{</div>}}}

\item{classify}{Variable to make predictions on}

\item{link_fun}{Something}

\item{tmb_component}{Something}

\item{tmb_type}{Something}

\item{factor_combine}{Something}

\item{trt_col_label}{Something}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Something
}\if{html}{\out{</div>}}
}
\description{
This function returns the predictions and confidence intervals of an
\code{asreml} or \code{glmmTMB} object.
}
