% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tables.R
\name{report_tables}
\alias{report_tables}
\title{A prediction function for models with more than one variable to predict on}
\usage{
report_tables(model, classify)
}
\arguments{
\item{model}{The model to generate predictions for.

The value may be:
\itemize{
\item \code{asreml}
\item \code{glmmTBM} (not yet implemented)
}}

\item{classify}{\code{character}
A string specifying the factors in the model to predict on. If multiple
are specified separate with either \code{*} or \code{:}. For example
\code{Nitrogen:Variety} or \code{Nitrogen*Variety}.}
}
\value{
\code{list} of \code{data.frame}
A list of data frames. The first data frame is the ANOVA for the model.
The remaining data frames are the prediction tables from the classify
object.
}
\description{
This function outputs the predictions for each factor for a single model.
Each factor is put on a separate Excel sheet.
}
\examples{
\dontshow{if (requireNamespace("asreml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(asreml)
model <- asreml(
    fixed = yield ~ Variety + Nitrogen + Variety:Nitrogen,
    random = ~idv(Blocks) + idv(Blocks):idv(Wplots),
    residual = ~idv(units),
    data = oats
)

report_tables(
    model,
    classify = "Variety:Nitrogen"
)

\dontrun{
# Using it to write with writexl
tables <- excel_prediction_file(
    model,
    classify = "Variety:Nitrogen",
)

writexl::write_xlsx(x = tables, path = "Prediction_Tables.xlsx")
}
\dontshow{\}) # examplesIf}
}
